--- app/services/bitflyer/auth.rb
@@ -1,64 +1,0 @@
-# frozen_string_literal: true
-
-module Bitflyer
-  class Auth < BaseApiClient
-    include Initializable
-
-    # For more information about how these URI templates are structured read the explaination in the RFC
-    # https://github.com/sporkmonger/addressable
-    # https://www.rfc-editor.org/rfc/rfc6570.txt
-    PATH = Addressable::Template.new("/auth{/segments*}{?query*}")
-    AUTHORIZATION_CODE = "authorization_code"
-
-    attr_reader :api_authorization_header
-    attr_accessor :access_token, :refresh_token, :expires_in
-
-    # Public: Crafts an authorization_url
-    #
-    # state - The state token which is passed in. This should be verified later
-    # redirect_uri - The url which Bitflyer should redirect to once the connected
-    #
-    # Returns an authorization url for establishing OAuth credentials
-    def self.authorize_url(state:, redirect_uri:)
-      query = {
-        client_id: Bitflyer.client_id,
-        response_type: Bitflyer.response_type,
-        scope: Bitflyer.scope,
-        state: state,
-        redirect_uri: redirect_uri,
-      }
-
-      authorization_path = PATH.expand(query: query, segments: nil)
-
-      "#{Bitflyer.oauth_uri}#{authorization_path}"
-    end
-
-    # Public: Calls the Bitflyer API for accessing an access_token
-    #
-    # code - The code obtained from Bitflyer during the initial authorization flow
-    #
-    # Returns an Auth Object
-    def self.token(code:, redirect_uri:)
-      Auth.new.token(code: code, redirect_uri: redirect_uri)
-    end
-
-    # Temporarily keep a method on the class instance for API Requests until the following issue is resolved
-    # https://github.com/brave-intl/publishers/issues/2779
-    def token(code:, redirect_uri:)
-      body = {
-        client_id: Bitflyer.client_id,
-        client_secret: Bitflyer.client_secret,
-        grant_type: AUTHORIZATION_CODE,
-        code: code,
-        redirect_uri: redirect_uri,
-      }
-      response = post(PATH.expand(segments: 'token'), body)
-
-      Auth.new(JSON.parse(response.body))
-    end
-
-    def api_base_uri
-      Bitflyer.api_base_uri
-    end
-  end
-end
